Aug 20 Commit 1 Restructure classes and files for CTC batching
