+++ data_util.py
@@ -1,0 +1,78 @@
+# !/usr/local/bin/python
+import numpy as np
+
+def sparse_tuple_from(sequences, dtype=np.int32):
+    """Create a sparse representention of x.
+    Args:
+        sequences: a list of lists of type dtype where each element is a sequence
+    Returns:
+        A tuple with (indices, values, shape)
+    """
+    indices = []
+    values = []
+
+    for n, seq in enumerate(sequences):
+        indices.extend(zip([n]*len(seq), range(len(seq))))
+        values.extend(seq)
+
+    indices = np.asarray(indices, dtype=np.int64)
+    values = np.asarray(values, dtype=dtype)
+    shape = np.asarray([len(sequences), np.asarray(indices).max(0)[1]+1], dtype=np.int64)
+
+    return indices, values, shape
+
+def load_dir(fp):
+    """
+    Summary:
+        Load raw paths data into arrays
+    Parameters:
+        fp : string path to data
+    Output:
+        Returns array of loaded files
+        loaded[0] = sound
+        loaded[1] = phonemes
+        loaded[2] = words
+        loaded[3] = text
+    """
+    with tf.name_scope('raw_data'):
+        ind = 0
+        raw_audio = []
+        phonemes = []
+        words = []
+        text = []
+        for __file in glob.iglob(path + '/*.*'):
+                if not ("SA" in __file):
+                    ind+=1
+                    if (ind%500==0):
+                        print(".", end="")
+                    if (".wav" in __file):
+                        raw_audio.append(__file)
+                    if (".PHN" in __file):
+                        with open(__file) as f:
+                            tmp_phn_file = []
+                            for line in f:
+                                tmp_phn = line.split(" ")
+                                tmp_phn_file.append(tmp_phn)
+                        phonemes.append(tmp_phn_file)
+                    if (".WRD" in __file):
+                        with open(__file) as f:
+                            tmp_wrd_file = []
+                            for line in f:
+                                tmp_wrd = line.split(" ")
+                                tmp_wrd_file.append(tmp_wrd)
+                        words.append(tmp_wrd_file)
+                    if (".TXT" in __file):
+                        with open(__file) as f:
+                            for line in f:
+                                res = ''.join([i for i in line if not i.isdigit()])
+                        text.append(res)
+
+        return raw_audio,phonemes,words,text,ind
+
+def pad_seq():
+
+
+def getData(filepath, minibatch_size, batches ,num_features ,num_classes):
+    """
+    Return training set and validation set (np arrays)
+    """
