+++ tf_model.py
@@ -9,9 +9,7 @@ import string
 import numpy as np
 import librosa.display as lib_disp
 import librosa.feature as lib_feat
-
-import tf_feed
-
+import data_util
 os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
 
 #PARAMS
@@ -20,59 +18,15 @@ path = '/home/jacky/Desktop/Spyre/__data/TIMIT/*/*/*'
 num_mfccs = 13
 batchsize = 10
 preprocess = 1
-learning_rate = 0.001
-window_cutoff = 5
+max_stepsize = 5
 sample_rate = 32000
-#0th indice +  space + blank label = 28 characters
-num_classes = ord('z') - ord('a') + 3
+num_hidden = 50
+learning_rate = 1e-2
+momentum = 0.9
+#0th indice + End indice + space + blank label = 28 characters
+num_classes = ord('z') - ord('a') + 4
 
-def load_dir(fp):
-    """
-    Summary:
-        Load raw paths data into arrays
-    Parameters:
-        fp : string path to data
-    Output:
-        Returns array of loaded files
-        loaded[0] = sound
-        loaded[1] = phonemes
-        loaded[2] = words
-        loaded[3] = text
-    """
-    with tf.name_scope('raw_data'):
-        ind = 0
-        raw_audio = []
-        phonemes = []
-        words = []
-        text = []
-        for __file in glob.iglob(path + '/*.*'):
-                if not ("SA" in __file):
-                    ind+=1
-                    if (ind%500==0):
-                        print(".", end="")
-                    if (".wav" in __file):
-                        raw_audio.append(__file)
-                    if (".PHN" in __file):
-                        with open(__file) as f:
-                            tmp_phn_file = []
-                            for line in f:
-                                tmp_phn = line.split(" ")
-                                tmp_phn_file.append(tmp_phn)
-                        phonemes.append(tmp_phn_file)
-                    if (".WRD" in __file):
-                        with open(__file) as f:
-                            tmp_wrd_file = []
-                            for line in f:
-                                tmp_wrd = line.split(" ")
-                                tmp_wrd_file.append(tmp_wrd)
-                        words.append(tmp_wrd_file)
-                    if (".TXT" in __file):
-                        with open(__file) as f:
-                            for line in f:
-                                res = ''.join([i for i in line if not i.isdigit()])
-                        text.append(res)
-
-        return raw_audio,phonemes,words,text
+inputs, labels = data_util.getData(num_examples, num_features, num_classes - 1)
 
 def features(rawsnd, num) :
     """
@@ -82,7 +36,7 @@ def features(rawsnd, num) :
         rawsnd : array with string paths to .wav files
         num : numbers of mfccs to compute
     Output:
-        Return a (num+28,window_cutoff*32)-dimensional Tensorflow feature vector, and length in
+        Return a (num+28,max_stepsize*32)-dimensional Tensorflow feature vector, and length in
         *num Amount of Mel Frequency Ceptral Coefficients
         *12 Chromagrams
         *7 bands of Spectral Contrast
@@ -91,7 +45,7 @@ def features(rawsnd, num) :
         *Spectral Rolloff
         *Spectral Centroid
     """
-    x, _ = librosa.load(rawsnd,sr=sample_rate, duration=window_cutoff)
+    x, _ = librosa.load(rawsnd,sr=sample_rate, duration=max_stepsize)
     s_tft = np.abs(librosa.stft(x))
     ft = lib_feat.mfcc(y=x, sr=sample_rate, n_mfcc=num)
     ft = np.append(ft,lib_feat.chroma_stft(S=s_tft, sr=sample_rate),axis=0)
@@ -100,29 +54,35 @@ def features(rawsnd, num) :
     ft = np.append(ft,lib_feat.zero_crossing_rate(y=x),axis=0)
     ft = np.append(ft,lib_feat.spectral_rolloff(y=x,sr=sample_rate),axis=0)
     ft = np.append(ft,lib_feat.spectral_centroid(y=x,sr=sample_rate),axis=0)
-    z = np.zeros((num+12+7+6+3,(window_cutoff*((sample_rate/1000)*2))-ft.shape[1]))
+    z = np.zeros((num+12+7+6+3,(max_stepsize*((sample_rate/1000)*2))-ft.shape[1]))
     ft = np.concatenate((ft,z),axis=1)
     #print(ft[13].astype(np.float16))
     return (ft)
 
-def preprocess(rawsnd) :
+def preprocess(rawsnd,stdev) :
     """
     #INCOMPLETE
-    If preprocess == 1, add additional white noise and
-    random sample from Urban8K
+    If preprocess == 1, add additional white noise with stdev (default 0.6)
     """
 
-def BLSTM():
+def BiRNN(x, weights, biases):
     """
-    #INCOMPLETE
-    #Class
+    #Prepare data shape to match `bidirectional_rnn` function requirements
+    # Current data input shape: (batch_size, n_steps, n_input)
+    # Required shape: 'n_steps' tensors list of shape (batch_size, n_input)
     """
+    # Unstack to get a list of 'n_steps' tensors of shape (batch_size, n_input)
+    x = tf.unstack(x, n_steps, 1)
 
-def train():
-    """
-    #INCOMPLETE
-    Trains the model
-    """
+    # Define lstm cells with tensorflow
+    # Forward direction cell
+    lstm_fw_cell = rnn.BasicLSTMCell(n_hidden, forget_bias=1.0)
+    # Backward direction cell
+    lstm_bw_cell = rnn.BasicLSTMCell(n_hidden, forget_bias=1.0)
+    outputs = rnn.static_bidirectional_rnn(lstm_fw_cell, lstm_bw_cell, x, dtype=tf.float32)
+
+    # Linear activation, using rnn inner loop last output
+    return tf.matmul(outputs[-1], weights['out']) + biases['out']
 
 init = tf.global_variables_initializer()
